requirements.txt
openai==1.17.0
python-dotenv==1.0.1
streamlit==1.33.0
scipy==1.13.0